Reorganize functions into proper files.
