+++ pythonx/jupyter_util.py
@@ -143,116 +143,3 @@ def find_signals():
                if v.startswith('SIG') and not v.startswith('SIG_')]
     signals.sort()
     return signals
-
-# -----------------------------------------------------------------------------
-#        Parsers
-# -----------------------------------------------------------------------------
-
-def parse_iopub_for_reply(msgs, line_number):
-    """Get kernel response from message pool (Async)
-    Param: line_number: the message number of the corresponding code
-    Use: some kernel (iperl) do not discriminate when clien ask user_expressions.
-        But still they give a printable output
-    """
-    res = -1
-
-    # Parse all execute
-    for msg in msgs:
-        # Get the result of execution
-        # 1 content
-        content = msg.get('content', False)
-        if not content: continue
-
-        # 2 execution _count
-        ec = int(content.get('execution_count', 0))
-        if not ec: continue
-        if line_number not in (-1, ec): continue
-
-        # 3 message type
-        msg_type = msg.get('header', {}).get('msg_type', '')
-        if msg_type not in ('execute_result', 'stream'): continue
-
-        # 4 text
-        res = content.get('data', {}).get('text/plain', -1)
-        res = res if res != -1 else content.get('text', -1)  # Jupyter bash style ...
-        break
-    return res
-
-
-def parse_messages(section_info, msgs):
-    """Message handler for Jupyter protocol (Async)
-
-    Takes all messages on the I/O Public channel, including stdout, stderr,
-    etc.
-    Returns: a list of the formatted strings of their content.
-
-    See also: <http://jupyter-client.readthedocs.io/en/stable/messaging.html>
-    """
-    # pylint: disable=too-many-branches
-    # TODO section_info is not perfectly async
-    # TODO remove complexity
-    res = []
-    for msg in msgs:
-        s = ''
-        default_count = section_info.monitor.cmd_count
-        if 'msg_type' not in msg['header']:
-            continue
-        msg_type = msg['header']['msg_type']
-
-        if msg_type == 'status':
-            # I don't care status (idle or busy)
-            continue
-
-        if msg_type == 'stream':
-            # Get data
-            text = strip_color_escapes(msg['content']['text'])
-            line_number = msg['content'].get('execution_count', default_count)
-            # Set prompt
-            if msg['content'].get('name', 'stdout') == 'stderr':
-                prompt = 'Err[{:d}]: '.format(line_number)
-                dots = (' ' * (len(prompt.rstrip()) - 4)) + '...x '
-            else:
-                prompt = 'Out[{:d}]: '.format(line_number)
-                dots = (' ' * (len(prompt.rstrip()) - 4)) + '...< '
-            s = prompt
-            # Add continuation line, if necessary
-            s += text.rstrip().replace('\n', '\n' + dots)
-            # Set cmd_count: if it changed
-            if line_number != default_count:
-                section_info.monitor.set_cmd_count(line_number)
-
-        elif msg_type == 'display_data':
-            s += msg['content']['data']['text/plain']
-
-        elif msg_type in ('execute_input', 'pyin'):
-            line_number = msg['content'].get('execution_count', default_count)
-            cmd = msg['content']['code']
-            s = prettify_execute_intput(line_number, cmd, section_info.lang.prompt_in)
-            # Set cmd_count: if it changed
-            if line_number != default_count:
-                section_info.monitor.set_cmd_count(line_number)
-
-        elif msg_type in ('execute_result', 'pyout'):
-            # Get the output
-            line_number = msg['content'].get('execution_count', default_count)
-            s = section_info.lang.prompt_out.format(line_number)
-            s += msg['content']['data']['text/plain']
-            # Set cmd_count: if it changed
-            if line_number != default_count:
-                section_info.monitor.set_cmd_count(line_number)
-
-        elif msg_type in ('error', 'pyerr'):
-            s = "\n".join(map(strip_color_escapes, msg['content']['traceback']))
-
-        elif msg_type == 'input_request':
-            section_info.vim.thread_echom('python input not supported in vim.', style='Error')
-            continue  # unsure what to do here... maybe just return False?
-
-        else:
-            section_info.vim.thread_echom("Message type {} unrecognized!".format(msg_type))
-            continue
-
-        # List all messages
-        res.append(s)
-
-    return res
